requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==3.2.12
django-allauth==0.48.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.0.1
PyJWT==2.3.0
python-telegram-bot==13.11
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
whitenoise==6.0.0
